refactor(types): reorganise internal/types into single-themed files (Pass 1)
Split the two grab-bag files (state.go ~1000 lines, types.go ~330) into
domain-themed files with zero externally-visible change: the package's
exported API surface and JSON/YAML serialization are byte-for-byte
identical. Verified during development with a temporary no-op suite
(exported-API snapshot + file-format round-trip goldens) that passed
both before and after the reorg, then removed.

- state.go: 1015 -> 261 lines (root State + orchestration/query methods)
- types.go -> auth.go, terraform.go, wizard_requests.go
- discovery split: discovery_msk.go, discovery_osk.go, discovery_common.go
- metrics.go, costs.go, schema_registry.go, processed.go extracted from state.go
- migration_{state,config,constants}.go collapsed into migration.go
  (+ MigrationType/Manifest moved from types.go)
- msk_credentials.go/osk_credentials.go -> credentials_msk.go/credentials_osk.go
  (pure rename; credentials dedup deferred to a later PR)
- self_managed_connect.go: merge logic + setter moved to discovery_common.go
  (co-located with the KafkaAdminClientInformation receiver type)

No behaviour change. Pass 2 (moving owner-specific types into their
consumer service packages) is a separate follow-up.
