--- internal/types/migration_config.go
@@ -1,39 +1,0 @@
-package types
-
-// MigrationConfig holds all domain configuration for a migration
-// This is pure data with no behavior - just fields that get serialized
-type MigrationConfig struct {
-	MigrationId  string `json:"migration_id"`
-	CurrentState string `json:"current_state"`
-
-	// Gateway configuration
-	KubeConfigPath string `json:"kube_config_path"`
-
-	// Source cluster configuration
-	SourceBootstrap string `json:"source_bootstrap"`
-
-	// Destination cluster configuration
-	ClusterBootstrap    string   `json:"cluster_bootstrap"`
-	ClusterId           string   `json:"cluster_id"`
-	ClusterRestEndpoint string   `json:"cluster_rest_endpoint"`
-	ClusterLinkName     string   `json:"cluster_link_name"`
-	Topics              []string `json:"topics"`
-
-	// Migration runtime data (populated during initialization)
-	ClusterLinkTopics  []string          `json:"cluster_link_topics"`
-	ClusterLinkConfigs map[string]string `json:"cluster_link_configs"`
-
-	// Operator intent: pause cluster-link consumer offset sync for the duration of execute.
-	// PauseConsumerOffsetSync records the operator's choice at init time.
-	// PauseConsumerOffsetSyncFlipped is set when kcp has executed the disable AlterConfigs and
-	// not yet restored — supports drift detection, idempotent resume, and remediation messaging.
-	PauseConsumerOffsetSync        bool `json:"pause_consumer_offset_sync"`
-	PauseConsumerOffsetSyncFlipped bool `json:"pause_consumer_offset_sync_flipped"`
-
-	// Gateway CR configuration
-	InitialCrName    string `json:"initial_cr_name"`
-	K8sNamespace     string `json:"k8s_namespace"`
-	InitialCrYAML    []byte `json:"initial_cr_yaml"`
-	FencedCrYAML     []byte `json:"fenced_cr_yaml"`
-	SwitchoverCrYAML []byte `json:"switchover_cr_yaml"`
-}
