--- internal/types/migration_constants.go
@@ -1,20 +1,0 @@
-package types
-
-// FSM State constants
-const (
-	StateUninitialized = "uninitialized"
-	StateInitialized   = "initialized"
-	StateLagsOk        = "lags_ok"
-	StateFenced        = "fenced"
-	StatePromoted      = "promoted"
-	StateSwitched      = "switched"
-)
-
-// FSM Event constants
-const (
-	EventInitialize  = "initialize"
-	EventWaitForLags = "wait_for_lags"
-	EventFence       = "fence"
-	EventPromote     = "promote"
-	EventSwitch      = "switch"
-)
