+++ internal/types/self_managed_connect.go
@@ -30,49 +30,3 @@ type SelfManagedConnectors struct {
 	Connectors []SelfManagedConnector   `json:"connectors"`
 	Metrics    *ProcessedClusterMetrics `json:"metrics,omitempty"`
 }
-
-// mergeSelfManagedConnectors merges connectors, with new taking precedence for duplicates (by name)
-func mergeSelfManagedConnectors(newConnectors, oldConnectors *SelfManagedConnectors) *SelfManagedConnectors {
-	if oldConnectors == nil || len(oldConnectors.Connectors) == 0 {
-		return newConnectors
-	}
-	if newConnectors == nil || len(newConnectors.Connectors) == 0 {
-		return oldConnectors
-	}
-
-	connectorsByName := make(map[string]SelfManagedConnector)
-	for _, c := range oldConnectors.Connectors {
-		connectorsByName[c.Name] = c
-	}
-	for _, c := range newConnectors.Connectors {
-		connectorsByName[c.Name] = c // new takes precedence
-	}
-
-	merged := make([]SelfManagedConnector, 0, len(connectorsByName))
-	for _, c := range connectorsByName {
-		merged = append(merged, c)
-	}
-
-	result := &SelfManagedConnectors{Connectors: merged}
-
-	// Preserve metrics: prefer new, fall back to old
-	if newConnectors.Metrics != nil {
-		result.Metrics = newConnectors.Metrics
-	} else if oldConnectors.Metrics != nil {
-		result.Metrics = oldConnectors.Metrics
-	}
-
-	return result
-}
-
-func (c *KafkaAdminClientInformation) SetSelfManagedConnectors(connectors []SelfManagedConnector) {
-	// Preserve existing metrics when updating connectors
-	var existingMetrics *ProcessedClusterMetrics
-	if c.SelfManagedConnectors != nil {
-		existingMetrics = c.SelfManagedConnectors.Metrics
-	}
-	c.SelfManagedConnectors = &SelfManagedConnectors{
-		Connectors: connectors,
-		Metrics:    existingMetrics,
-	}
-}
