+++ internal/types/wizard_requests.go
@@ -1,122 +1,5 @@
 package types
 
-import (
-	"fmt"
-	"strconv"
-)
-
-type TerraformState struct {
-	Outputs TerraformOutputOld `json:"outputs"`
-}
-
-// a type for the output.json file in the target_env folder
-// NOTE: This will be deprecated once we are completely on the HCL-servrice based approach.
-type TerraformOutputOld struct {
-	ConfluentCloudClusterApiKey                TerraformOutputValue `json:"confluent_cloud_cluster_api_key"`
-	ConfluentCloudClusterApiKeySecret          TerraformOutputValue `json:"confluent_cloud_cluster_api_key_secret"`
-	ConfluentCloudClusterId                    TerraformOutputValue `json:"confluent_cloud_cluster_id"`
-	ConfluentCloudClusterRestEndpoint          TerraformOutputValue `json:"confluent_cloud_cluster_rest_endpoint"`
-	ConfluentCloudClusterBootstrapEndpoint     TerraformOutputValue `json:"confluent_cloud_cluster_bootstrap_endpoint"`
-	ConfluentPlatformControllerBootstrapServer TerraformOutputValue `json:"confluent_platform_controller_bootstrap_server"`
-}
-
-type TerraformOutputValue struct {
-	Sensitive bool   `json:"sensitive"`
-	Type      string `json:"type"`
-	Value     any    `json:"value"`
-}
-
-// AuthType represents the different authentication types supported by MSK clusters
-type AuthType string
-
-const (
-	AuthTypeSASLSCRAM                AuthType = "SASL/SCRAM"
-	AuthTypeIAM                      AuthType = "SASL/IAM"
-	AuthTypeTLS                      AuthType = "TLS"
-	AuthTypeUnauthenticatedPlaintext AuthType = "Unauthenticated (Plaintext)"
-	AuthTypeUnauthenticatedTLS       AuthType = "Unauthenticated (TLS Encryption)"
-	AuthTypeSASLPlain                AuthType = "SASL/PLAIN"
-)
-
-// SchemaRegistryAuthType represents the different authentication types supported by Schema Registry
-type SchemaRegistryAuthType string
-
-const (
-	SchemaRegistryAuthTypeUnauthenticated SchemaRegistryAuthType = "Unauthenticated"
-	SchemaRegistryAuthTypeBasicAuth       SchemaRegistryAuthType = "BasicAuth"
-)
-
-func (a AuthType) IsValid() bool {
-	switch a {
-	case AuthTypeSASLSCRAM, AuthTypeIAM, AuthTypeTLS, AuthTypeUnauthenticatedPlaintext, AuthTypeUnauthenticatedTLS, AuthTypeSASLPlain:
-		return true
-	default:
-		return false
-	}
-}
-
-// Values returns all possible AuthType values as strings
-func (a AuthType) Values() []string {
-	return AllAuthTypes()
-}
-
-// AllAuthTypes returns all possible AuthType values as strings
-// This can be called statically without needing an AuthType instance
-func AllAuthTypes() []string {
-	return []string{
-		string(AuthTypeSASLSCRAM),
-		string(AuthTypeIAM),
-		string(AuthTypeTLS),
-		string(AuthTypeUnauthenticatedPlaintext),
-		string(AuthTypeUnauthenticatedTLS),
-		string(AuthTypeSASLPlain),
-	}
-}
-
-type MigrationType int
-
-const (
-	PublicMskEndpoints                   MigrationType = 1
-	ExternalOutboundClusterLink          MigrationType = 2
-	ExternalOutboundClusterLinkPlaintext MigrationType = 3
-	JumpClusterSaslScram                 MigrationType = 4
-	JumpClusterIam                       MigrationType = 5
-)
-
-func (m MigrationType) IsValid() bool {
-	switch m {
-	case PublicMskEndpoints, ExternalOutboundClusterLink, ExternalOutboundClusterLinkPlaintext, JumpClusterSaslScram, JumpClusterIam:
-		return true
-	default:
-		return false
-	}
-}
-
-func (m MigrationType) RequiresSaslScram() bool {
-	switch m {
-	case PublicMskEndpoints, ExternalOutboundClusterLink, JumpClusterSaslScram:
-		return true
-	default:
-		return false
-	}
-}
-
-func ToMigrationType(input string) (MigrationType, error) {
-	value, err := strconv.Atoi(input)
-	if err != nil {
-		return 0, fmt.Errorf("invalid input: must be a number")
-	}
-	m := MigrationType(value)
-	if !m.IsValid() {
-		return 0, fmt.Errorf("invalid MigrationType value: %d", value)
-	}
-	return m, nil
-}
-
-type Manifest struct {
-	MigrationInfraType MigrationType `json:"migration_infra_type"`
-}
-
 type TargetClusterWizardRequest struct {
 	AwsRegion              string   `json:"aws_region"`
 	NeedsEnvironment       bool     `json:"needs_environment"`
@@ -132,29 +15,6 @@ type TargetClusterWizardRequest struct {
 	PreventDestroy         bool     `json:"prevent_destroy"`
 	VpcId                  string   `json:"vpc_id"`
 	SubnetCidrRanges       []string `json:"subnet_cidr_ranges"`
-}
-
-type TerraformFiles struct {
-	MainTf           string            `json:"main.tf"`
-	ProvidersTf      string            `json:"providers.tf"`
-	VariablesTf      string            `json:"variables.tf"`
-	InputsAutoTfvars string            `json:"inputs.auto.tfvars"`
-	OutputsTf        string            `json:"outputs.tf"`
-	PerPrincipalTf   map[string]string `json:"per_principal_tf,omitempty"`
-}
-
-type TerraformVariable struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-	Sensitive   bool   `json:"sensitive"`
-	Type        string `json:"type"`
-}
-
-type TerraformOutput struct {
-	Name        string
-	Description string
-	Sensitive   bool
-	Value       string
 }
 
 type MigrationWizardRequest struct {
@@ -287,43 +147,4 @@ type GlueSchemaRegistryMigrationConfig struct {
 	RegistryName string       `json:"registry_name"`
 	Region       string       `json:"region"`
 	Schemas      []GlueSchema `json:"schemas"`
-}
-
-// MigrationInfraTerraformModule represents a Terraform module within the migration infrastructure
-// configuration. Each module contains its own Terraform files and additional assets.
-type MigrationInfraTerraformModule struct {
-	Name            string            `json:"name"`
-	MainTf          string            `json:"main.tf"`
-	VariablesTf     string            `json:"variables.tf"`
-	OutputsTf       string            `json:"outputs.tf"`
-	VersionsTf      string            `json:"versions.tf"`
-	AdditionalFiles map[string]string `json:"additional_files"`
-}
-
-// MigrationInfraTerraformProject represents the complete Terraform configuration for migration
-// infrastructure. "project" = root config + modules
-type MigrationInfraTerraformProject struct {
-	MainTf           string                          `json:"main.tf"`
-	ProvidersTf      string                          `json:"providers.tf"`
-	VariablesTf      string                          `json:"variables.tf"`
-	OutputsTf        string                          `json:"outputs.tf"`
-	ReadmeMd         string                          `json:"README.md"`
-	InputsAutoTfvars string                          `json:"inputs.auto.tfvars"`
-	Modules          []MigrationInfraTerraformModule `json:"modules"`
-}
-
-// MigrationScriptsTerraformProject represents the complete Terraform configuration for migration scripts
-type MigrationScriptsTerraformProject struct {
-	// not really a module, but its the same structure
-	Folders []MigrationScriptsTerraformFolder `json:"modules"`
-}
-
-// MigrationScriptsTerraformFolder represents a Terraform folder within the migration scripts
-type MigrationScriptsTerraformFolder struct {
-	Name             string            `json:"name"`
-	MainTf           string            `json:"main.tf"`
-	ProvidersTf      string            `json:"providers.tf"`
-	VariablesTf      string            `json:"variables.tf"`
-	InputsAutoTfvars string            `json:"inputs.auto.tfvars"`
-	AdditionalFiles  map[string]string `json:"additional_files,omitempty"`
 }
